+++ src/components/shared/sidebar/ProjectTabToolbarSidebar.tsx
@@ -1,6 +1,6 @@
 import type { ReactNode } from 'react';
 import { Link } from 'react-router-dom';
-import { SidebarButton } from '../../primitives';
+import { SidebarButton } from './SidebarButton';
 import { SidebarButtonGroup } from './SidebarButtonGroup';
 
 interface ProjectTabToolbarSidebarProps {