+++ src/components/shared/sidebar/SidebarButton.tsx
@@ -8,7 +8,7 @@ import {
   type ReactElement,
   type Ref,
 } from 'react';
-import { cn } from '../../lib/utils';
+import { cn } from '../../../lib/utils';
 
 type SidebarButtonVariant = 'default' | 'add';