+++ src/components/shared/sidebar/index.ts
@@ -1,2 +1,3 @@
 export { ProjectTabToolbarSidebar } from './ProjectTabToolbarSidebar';
 export { SidebarButtonGroup } from './SidebarButtonGroup';
+export { SidebarButton } from './SidebarButton';